requirements.txt
Flask>=3.0.0
Flask-Limiter>=3.7.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
joblib>=1.3.2
idna>=3.6
tldextract>=5.1.2
rapidfuzz>=3.6.1
requests>=2.31.0
reportlab>=4.1.0
transformers>=4.45.0
torch>=2.2.0
